Extract territory-lookup
